Fixed to not re-register if the number of circles is the same

+++ db/seeds/004_circles.rb
@@ -2,10 +2,13 @@
 
 require 'csv'
 
+circles = CSV.table('db/fixtures/circles.tsv', col_sep: "\t", converters: nil)
+return if circles.size == Circle.count
+
 ActiveRecord::Base.connection.execute('TRUNCATE TABLE circles_albums, circles;')
 insert_data = []
 now = Time.zone.now
-CSV.table('db/fixtures/circles.tsv', col_sep: "\t", converters: nil).each do |o|
+circles.each do |o|
   insert_data << {
     name: o[:name],
     created_at: now,
